Asteroid temp (#11)
A few changes were made:

- Create a windows CI/CD script that will compile with MVSC and Ninja
- Remove the docker usage from ubuntu compilation action, as they take too much to compile. This reduces the test from 7 min down to 4 min.
- Add .stfolder, use it instead of changing computer by commiting
- Remove the need to use NormalEntity class. This is because I got tired to maintain 1 billion classes equivalent to Scene.h. In the future I could join Systems too.
- This effect also mean that new objects can be created as pure Entities, no need to make classes if you don't want. Instead, you can create classes when making the script. Which makes sense since scripts are meant to fill the role of OOP
- Passing DeltaTime for onRender and onUpdate
- DefaultTexture enum fields had to be renamed, because windows in a bitch and keep declaring the conflicting MACROS
- Pedantic options won't generate more errors, because I clear them all
- Sweep And Prune Algorithm was added
- Union inside Sprite.h was replaced with std::variant, which has safety checks. Although I don't know if this increase cpu or memory overhead
- UNUSED macro for suppressing unused argument warnings
- Some bugs in the reactions were eliminated

+++ Pain/include/ECS/Registry/Archetype.h
@@ -41,7 +41,7 @@ public:
   template <typename C, typename... Args> Column pushComponent(Args &&...args)
   {
     std::vector<C> &componentArray = createComponent<C>();
-    Column index = Column(componentArray.size());
+    Column index = Column(static_cast<int>(componentArray.size()));
     componentArray.emplace_back(std::forward<Args>(args)...);
     return index;
   }
@@ -50,7 +50,7 @@ public:
   // remove
   // ---------------------------------------------------- //
 
-  template <typename... Components> std::uint32_t remove(Column column)
+  template <typename... Components> std::int32_t remove(Column column)
   {
     static_assert(sizeof...(Components) != 0,
                   "Not allowed to remove from empty archetypes");
@@ -91,7 +91,7 @@ public:
     P_ASSERT((unsigned)entityIndex <= componentArray.size(),
              "Entity index {} out of range on array of size {}",
              entityIndex.value, componentArray.size());
-    return componentArray[entityIndex];
+    return componentArray[(unsigned)entityIndex];
   }
 
   template <typename T> T &fetchComponent(Column entityIndex)
@@ -100,7 +100,7 @@ public:
     P_ASSERT((unsigned)entityIndex <= componentArray.size(),
              "Entity index {} out of range on array of size {}",
              entityIndex.value, componentArray.size());
-    return componentArray[entityIndex];
+    return componentArray[(unsigned)entityIndex];
   }
 
   template <typename C> const std::vector<C> &getComponent() const
